backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
google-auth==2.25.2
google-generativeai==0.3.2
requests==2.31.0

